+++ huey/tests/__init__.py
@@ -5,6 +5,7 @@ from huey.tests.test_crontab import *
 from huey.tests.test_django_tasks import *
 from huey.tests.test_helpers import *
 from huey.tests.test_immediate import *
+from huey.tests.test_migration import *
 from huey.tests.test_mini import *
 from huey.tests.test_postgres import *
 from huey.tests.test_priority import *